refactor: down arrow 제거후, 공통 arrow로 대체
