+++ client/src/icons/SmallArrow.tsx
@@ -15,3 +15,14 @@ export function SmallRightArrow() {
     </Icon>
   );
 }
+
+export function SmallDownArrow() {
+  return (
+    <Icon width='8' height='6' viewBox='0 0 8 6' fill='none' xmlns='http://www.w3.org/2000/svg'>
+      <path
+        d='M3.38753 5.51646C2.41988 4.42493 1.25968 3.06299 0.372903 2.01202C-0.17385 1.36403 0.289157 0.376953 1.13699 0.376953L6.91084 0.376953C7.75117 0.376953 8.21677 1.35058 7.68925 2.0047L4.89114 5.47436C4.50771 5.94981 3.79272 5.97351 3.38753 5.51646Z'
+        fill='black'
+      />
+    </Icon>
+  );
+}
