+++ client/src/icons/types.ts
@@ -1,0 +1,5 @@
+export interface IconProps {
+  w?: number | string;
+  h?: number | string;
+  color?: string;
+}
